feat:create_shipment

+++ src/main/java/vn/ptit/service/user/DeleteUserService.java
@@ -1,6 +1,8 @@
 package vn.ptit.service.user;
 
+import lombok.SneakyThrows;
 import org.springframework.stereotype.Service;
+import vn.ptit.exception.DataNotFoundException;
 import vn.ptit.model.User;
 import vn.ptit.repository.user.IUserRepository;
 
@@ -12,8 +14,12 @@ public class DeleteUserService {
         this.userRepository = userRepository;
     }
 
+    @SneakyThrows
     public void delete(Long id){
         User user = userRepository.getById(id);
+        if(user == null){
+            throw new DataNotFoundException("User not found");
+        }
         user.delete();
         userRepository.save(user);
     }
